Fix NAD7050 hangs with identical settings
Fix NAD7050 hangs when setting source or power to the value it already has, as well as a hang when setting the source while the power is off. This normally only occurs in automations, because normally only commands are send when the status changes, but in an automation people can choose to just set things without first checking the current status.