Wait for the requested response
When sending a telnet request and at the same time
using the remote control the reponse was sometimes
incorrect. This solution is probably not 100%

Eg. sending main.volume? while fibbling with the remote
does not guarantee that the response is for the
main.volume? command, it could be a response
from remote control volume changes. But filtering
should work for e.g. main.power? while pressing
vol+/- on the remote

Tested within home-assistant