Introduce transports
The rs232 code and telnet are basically the same. The official NAD
documentation confirms this, the data sent must be identical for both
protocols.
Abstract the difference away in simple classes instead of having
duplicate code.

The TCP class looks sufficiently different, since I cannot test that, I
do not dare touching it.

This also fixes the need for a 0.5 second wait in the serial port
connection by rather reading until a '\r' is received.
Works more reliably for my receiver and I assume all models.

Add a first test to be able to verify different device's protocols.