Add fake NAD C356BE
This implements the complete RS232 API of the receiver in a mock class.
With this the receiver can be tested locally without having any serial port.

Add a test that checks that communication with that particular amplifier
doesn't break again. Similar tests are needed for other models, but
since this is the only amp I have, I'm not easily able to write them.

The sources are for now left out because the current implementation
assumes sources to be integers, while the V2 protocol is purely string
based.